refactored into proper package pyautoml
