+++ pyautoml/cleaning/categorical.py
@@ -8,7 +8,7 @@ ReplaceMissingRemoveRow
 import numpy as np
 import pandas as pd
 
-from data.util import DropAndReplaceColumns, _FunctionInputValidation
+from pyautoml.util import DropAndReplaceColumns, _FunctionInputValidation
 
 #TODO: Implement KNN, and replacing with most common category
